AggregatedAttribute should add both new and old values when preparing the local_condition.
Currently AggregatedAttribute only update the new object, so if we change the relationship from one to another. Only new object aggregated attribute will be updated and the old one remain unchanged. To fix this, we should add both the added and deleted value to the list when preparing the local_condition in AggregatedAttribute.